batch-scripts/requirements.txt
azure-batch==8.0.0
